Refactored agents
