--- agent/chirpy_agent.py
@@ -1,3 +1,0 @@
-from agent.agents.local.local_agent import LocalAgent
-
-chirpy_agent = LocalAgent()
